Commit #12: Fixing Data Bugs
Some things I didn't include in the last commit.

This commit addresses the issue with data being passed from this
simulator to OpenICE being inconsistent.

Most data instances from this simulator in the system explorer of
OpenICE would stick around after closing the program. Now those
instances properly unregister from the system and disappear from the
system explorer.

Data is only passed to the OpenICE supervisor if the flow rate (L/min)
is greater than 0.1. This will be reduced to 0 when the dial
implementation is changed to make it easier to rotate to full 0.0.

Time on bypass is only calculated when flow rate (L/min) is greater than
0.1. This will be reduced to 0 when the dial implementation is changed
to make it easier to rotate to full 0.0.